feat(InteriorLeftNav): default active href prop
Currently setting `activeHref` means that any time any prop updates, it resets the internal activeHref to the prop. Because of this, there is no way to set the initial default active href and then let the component's logic take over when items are clicked.
This add `defaultActiveHref` (similar to `value` vs `defaultValue`) to allow for unmanaged activeHref to be defaulted. It is mostly needed when `window.location.pathname` doesn't work (such as when links have hashes).